requirements.txt
wheel
Django~=2.0.6
requests
python-decouple
